printf文のフォーマット指定子に型キャストを追加

+++ src/reflector.c
@@ -378,7 +378,7 @@ int main(int argc, char *argv[])
 			const struct stamp_reflector_packet *packet =
 				(const struct stamp_reflector_packet *)buffer;
 			printf("Reflected packet Seq: %" PRIu32 " from %s:%d (TTL: %d)\n",
-				   ntohl(packet->sender_seq_num),
+				   (uint32_t)ntohl(packet->sender_seq_num),
 				   inet_ntoa(cliaddr.sin_addr),
 				   ntohs(cliaddr.sin_port),
 				   ttl);
